refactor: 앱 업데이트 토스트 로직을 useAppUpdateToast로 분리
